feat: Added gradient-show and form-show elements
The elements gradient-show and form-show are used to display the gradient image and form respectively.

Added a function computeGradientShowAndFormShowWidths which when called resizes the two newly added elements

The newly added function computeGradientShowAndFormShowWidths is called on DOMContentLoaded and whenever the window is resized, added callbacks to that effect.

Added a new file AbstractConstants which holds the scss constants large and medium used in media queries for usage in JS.

Basic styling added on main and gradient-show.