feat: added card component to model the cards found on design
Added templates for the card front and card back

Added the app-card-front and app-card-back components to gradient-show

gradient-show is position relative to accommodate its children cards.

Added a new base class Card for the app-card-front and app-card-back components, this does very little in terms of js code as of now and only adds the css class 'card' to this if super.connectedCallback is called.

Added two new components app-card-front and app-card-back along with its associated responsive styling.