feat: added a new web component app-date-input
Moved the input template into its own files.

Added code to test the working of date input

DateInput component is extended from the Input base component which is the parent for all types of input

Styling for any input component should be found in Input.scss

Input is the base class for all inputs. it provides the template as well as neat functions to set stuff like the error