feat: CardFront and CardBack now connected to form inputs
Fixed spelling in card-back-template to make it consistent with the project.

Moved card-front name placeholder to js as it is now accessed from there.

Added helper methods in CardFront and CardBack to display various fields.

Adjusted input max length in Name Input so that the display in CardFront doesn't overflow.